+++ tests/Geocoder/Tests/Provider/GeoIPsTest.php
@@ -135,8 +135,8 @@ class GeoIPsTest extends TestCase
         /** @var Location $result */
         $result = $results->first();
         $this->assertInstanceOf('\Geocoder\Model\Address', $result);
-        $this->assertNull($result->getCoordinates()->getLatitude());
-        $this->assertNull($result->getCoordinates()->getLongitude());
+        $this->assertNull($result->getCoordinates());
+
         $this->assertNull($result->getPostalCode());
         $this->assertNull($result->getLocality());
         $this->assertEmpty($result->getAdminLevels());