+++ tests/Geocoder/Tests/Provider/HostIpTest.php
@@ -55,8 +55,8 @@ class HostIpTest extends TestCase
         /** @var Location $result */
         $result = $results->first();
         $this->assertInstanceOf('\Geocoder\Model\Address', $result);
-        $this->assertNull($result->getCoordinates()->getLatitude());
-        $this->assertNull($result->getCoordinates()->getLongitude());
+        $this->assertNull($result->getCoordinates());
+
         $this->assertNull($result->getPostalCode());
         $this->assertNull($result->getTimezone());
         $this->assertEmpty($result->getAdminLevels());
@@ -146,7 +146,7 @@ class HostIpTest extends TestCase
         /** @var Location $result */
         $result = $results->first();
         $this->assertInstanceOf('\Geocoder\Model\Address', $result);
-        $this->assertNull($result->getCoordinates()->getLatitude());
-        $this->assertNull($result->getCoordinates()->getLongitude());
+        $this->assertNull($result->getCoordinates());
+
     }
 }