+++ tests/Geocoder/Tests/Provider/IpInfoDbTest.php
@@ -83,8 +83,8 @@ class IpInfoDbTest extends TestCase
         /** @var Location $result */
         $result = $results->first();
         $this->assertInstanceOf('\Geocoder\Model\Address', $result);
-        $this->assertNull($result->getCoordinates()->getLatitude());
-        $this->assertNull($result->getCoordinates()->getLongitude());
+        $this->assertNull($result->getCoordinates());
+
         $this->assertNull($result->getPostalCode());
         $this->assertNull($result->getTimezone());
         $this->assertEmpty($result->getAdminLevels());
@@ -181,8 +181,8 @@ class IpInfoDbTest extends TestCase
         /** @var Location $result */
         $result = $results->first();
         $this->assertInstanceOf('\Geocoder\Model\Address', $result);
-        $this->assertNull($result->getCoordinates()->getLatitude());
-        $this->assertNull($result->getCoordinates()->getLongitude());
+        $this->assertNull($result->getCoordinates());
+
         $this->assertNull($result->getPostalCode());
         $this->assertNull($result->getLocality());
         $this->assertEmpty($result->getAdminLevels());