+++ tests/Geocoder/Tests/Provider/MaxMindBinaryTest.php
@@ -56,7 +56,7 @@ class MaxMindBinaryTest extends TestCase
 
         $this->assertEquals('43.089200000000005', $result->getCoordinates()->getLatitude(), '', 0.001);
         $this->assertEquals('-76.025000000000006', $result->getCoordinates()->getLongitude(), '', 0.001);
-        $this->assertFalse($result->getBounds()->isDefined());
+        $this->assertNull($result->getBounds());
         $this->assertNull($result->getStreetNumber());
         $this->assertNull($result->getStreetName());
         $this->assertNull($result->getPostalCode());
@@ -84,7 +84,7 @@ class MaxMindBinaryTest extends TestCase
 
         $this->assertEquals('41.543299999999988', $result->getCoordinates()->getLatitude(), '', 0.001);
         $this->assertEquals('2.1093999999999937', $result->getCoordinates()->getLongitude(), '', 0.001);
-        $this->assertFalse($result->getBounds()->isDefined());
+        $this->assertNull($result->getBounds());
         $this->assertNull($result->getStreetNumber());
         $this->assertNull($result->getStreetName());
         $this->assertNull($result->getPostalCode());