+++ tests/Geocoder/Tests/Provider/NominatimTest.php
@@ -21,7 +21,7 @@ class NominatimTest extends TestCase
         $this->assertInstanceOf('\Geocoder\Model\Address', $result);
         $this->assertEquals(48.8565056, $result->getCoordinates()->getLatitude(), '', 0.01);
         $this->assertEquals(2.3521334, $result->getCoordinates()->getLongitude(), '', 0.01);
-        $this->assertTrue($result->getBounds()->isDefined());
+        $this->assertNotNull($result->getBounds());
         $this->assertEquals(48.8155250549316, $result->getBounds()->getSouth(), '', 0.01);
         $this->assertEquals(2.22412180900574, $result->getBounds()->getWest(), '', 0.01);
         $this->assertEquals(48.902156829834, $result->getBounds()->getNorth(), '', 0.01);
@@ -43,7 +43,7 @@ class NominatimTest extends TestCase
         $this->assertInstanceOf('\Geocoder\Model\Address', $result);
         $this->assertEquals(48.8588408, $result->getCoordinates()->getLatitude(), '', 0.01);
         $this->assertEquals(2.32003465529896, $result->getCoordinates()->getLongitude(), '', 0.01);
-        $this->assertTrue($result->getBounds()->isDefined());
+        $this->assertNotNull($result->getBounds());
         $this->assertEquals(48.8155250549316, $result->getBounds()->getSouth(), '', 0.01);
         $this->assertEquals(2.22412180900574, $result->getBounds()->getWest(), '', 0.01);
         $this->assertEquals(48.902156829834, $result->getBounds()->getNorth(), '', 0.01);
@@ -65,7 +65,7 @@ class NominatimTest extends TestCase
         $this->assertInstanceOf('\Geocoder\Model\Address', $result);
         $this->assertEquals(35.28687645, $result->getCoordinates()->getLatitude(), '', 0.01);
         $this->assertEquals(-93.7354879210082, $result->getCoordinates()->getLongitude(), '', 0.01);
-        $this->assertTrue($result->getBounds()->isDefined());
+        $this->assertNotNull($result->getBounds());
         $this->assertEquals(35.2672462463379, $result->getBounds()->getSouth(), '', 0.01);
         $this->assertEquals(-93.7618103027344, $result->getBounds()->getWest(), '', 0.01);
         $this->assertEquals(35.3065032958984, $result->getBounds()->getNorth(), '', 0.01);
@@ -87,7 +87,7 @@ class NominatimTest extends TestCase
         $this->assertInstanceOf('\Geocoder\Model\Address', $result);
         $this->assertEquals(33.6751155, $result->getCoordinates()->getLatitude(), '', 0.01);
         $this->assertEquals(-95.5502662477703, $result->getCoordinates()->getLongitude(), '', 0.01);
-        $this->assertTrue($result->getBounds()->isDefined());
+        $this->assertNotNull($result->getBounds());
         $this->assertEquals(33.6118507385254, $result->getBounds()->getSouth(), '', 0.01);
         $this->assertEquals(-95.6279296875, $result->getBounds()->getWest(), '', 0.01);
         $this->assertEquals(33.7383804321289, $result->getBounds()->getNorth(), '', 0.01);
@@ -109,7 +109,7 @@ class NominatimTest extends TestCase
         $this->assertInstanceOf('\Geocoder\Model\Address', $result);
         $this->assertEquals(38.2097987, $result->getCoordinates()->getLatitude(), '', 0.01);
         $this->assertEquals(-84.2529869, $result->getCoordinates()->getLongitude(), '', 0.01);
-        $this->assertTrue($result->getBounds()->isDefined());
+        $this->assertNotNull($result->getBounds());
         $this->assertEquals(38.1649208068848, $result->getBounds()->getSouth(), '', 0.01);
         $this->assertEquals(-84.3073272705078, $result->getBounds()->getWest(), '', 0.01);
         $this->assertEquals(38.2382736206055, $result->getBounds()->getNorth(), '', 0.01);
@@ -140,7 +140,7 @@ class NominatimTest extends TestCase
         $this->assertInstanceOf('\Geocoder\Model\Address', $result);
         $this->assertEquals(45.7586841, $result->getCoordinates()->getLatitude(), '', 0.01);
         $this->assertEquals(3.1354075, $result->getCoordinates()->getLongitude(), '', 0.01);
-        $this->assertTrue($result->getBounds()->isDefined());
+        $this->assertNotNull($result->getBounds());
         $this->assertEquals(45.7576484680176, $result->getBounds()->getSouth(), '', 0.01);
         $this->assertEquals(3.13258004188538, $result->getBounds()->getWest(), '', 0.01);
         $this->assertEquals(45.7595367431641, $result->getBounds()->getNorth(), '', 0.01);
@@ -162,7 +162,7 @@ class NominatimTest extends TestCase
         $this->assertInstanceOf('\Geocoder\Model\Address', $result);
         $this->assertEquals(45.7586841, $result->getCoordinates()->getLatitude(), '', 0.01);
         $this->assertEquals(3.1354075, $result->getCoordinates()->getLongitude(), '', 0.01);
-        $this->assertTrue($result->getBounds()->isDefined());
+        $this->assertNotNull($result->getBounds());
         $this->assertEquals(45.7576484680176, $result->getBounds()->getSouth(), '', 0.01);
         $this->assertEquals(3.13258004188538, $result->getBounds()->getWest(), '', 0.01);
         $this->assertEquals(45.7595367431641, $result->getBounds()->getNorth(), '', 0.01);
@@ -193,7 +193,7 @@ class NominatimTest extends TestCase
         $this->assertInstanceOf('\Geocoder\Model\Address', $result);
         $this->assertEquals(60.4539, $result->getCoordinates()->getLatitude(), '', 0.001);
         $this->assertEquals(22.2568, $result->getCoordinates()->getLongitude(), '', 0.001);
-        $this->assertFalse($result->getBounds()->isDefined());
+        $this->assertNull($result->getBounds());
         $this->assertEquals(35, $result->getStreetNumber());
         $this->assertEquals('Läntinen Pitkäkatu', $result->getStreetName());
         $this->assertEquals(20100, $result->getPostalCode());
@@ -230,7 +230,7 @@ class NominatimTest extends TestCase
         $this->assertInstanceOf('\Geocoder\Model\Address', $result);
         $this->assertEquals(50.1856803, $result->getCoordinates()->getLatitude(), '', 0.01);
         $this->assertEquals(8.6506285, $result->getCoordinates()->getLongitude(), '', 0.01);
-        $this->assertTrue($result->getBounds()->isDefined());
+        $this->assertNotNull($result->getBounds());
         $this->assertEquals(50.1851196289062, $result->getBounds()->getSouth(), '', 0.01);
         $this->assertEquals(8.64984607696533, $result->getBounds()->getWest(), '', 0.01);
         $this->assertEquals(50.1860122680664, $result->getBounds()->getNorth(), '', 0.01);
@@ -251,7 +251,7 @@ class NominatimTest extends TestCase
         $this->assertInstanceOf('\Geocoder\Model\Address', $result);
         $this->assertEquals(50.1845911, $result->getCoordinates()->getLatitude(), '', 0.01);
         $this->assertEquals(8.6540194, $result->getCoordinates()->getLongitude(), '', 0.01);
-        $this->assertTrue($result->getBounds()->isDefined());
+        $this->assertNotNull($result->getBounds());
         $this->assertEquals(50.1840019226074, $result->getBounds()->getSouth(), '', 0.01);
         $this->assertEquals(8.65207481384277, $result->getBounds()->getWest(), '', 0.01);
         $this->assertEquals(50.1851234436035, $result->getBounds()->getNorth(), '', 0.01);
@@ -272,7 +272,7 @@ class NominatimTest extends TestCase
         $this->assertInstanceOf('\Geocoder\Model\Address', $result);
         $this->assertEquals(50.1862884, $result->getCoordinates()->getLatitude(), '', 0.01);
         $this->assertEquals(8.6493167, $result->getCoordinates()->getLongitude(), '', 0.01);
-        $this->assertTrue($result->getBounds()->isDefined());
+        $this->assertNotNull($result->getBounds());
         $this->assertEquals(50.1862106323242, $result->getBounds()->getSouth(), '', 0.01);
         $this->assertEquals(8.64931583404541, $result->getBounds()->getWest(), '', 0.01);
         $this->assertEquals(50.1862907409668, $result->getBounds()->getNorth(), '', 0.01);
@@ -293,7 +293,7 @@ class NominatimTest extends TestCase
         $this->assertInstanceOf('\Geocoder\Model\Address', $result);
         $this->assertEquals(50.1861344, $result->getCoordinates()->getLatitude(), '', 0.01);
         $this->assertEquals(8.649578, $result->getCoordinates()->getLongitude(), '', 0.01);
-        $this->assertTrue($result->getBounds()->isDefined());
+        $this->assertNotNull($result->getBounds());
         $this->assertEquals(50.1860084533691, $result->getBounds()->getSouth(), '', 0.01);
         $this->assertEquals(8.64943885803223, $result->getBounds()->getWest(), '', 0.01);
         $this->assertEquals(50.1862144470215, $result->getBounds()->getNorth(), '', 0.01);
@@ -349,7 +349,7 @@ class NominatimTest extends TestCase
         $this->assertInstanceOf('\Geocoder\Model\Address', $result);
         $this->assertEquals(43.6189768, $result->getCoordinates()->getLatitude(), '', 0.01);
         $this->assertEquals(1.4564493, $result->getCoordinates()->getLongitude(), '', 0.01);
-        $this->assertTrue($result->getBounds()->isDefined());
+        $this->assertNotNull($result->getBounds());
         $this->assertEquals(43.6159553527832, $result->getBounds()->getSouth(), '', 0.01);
         $this->assertEquals(1.45302963256836, $result->getBounds()->getWest(), '', 0.01);
         $this->assertEquals(43.623119354248, $result->getBounds()->getNorth(), '', 0.01);
@@ -380,7 +380,7 @@ class NominatimTest extends TestCase
         $this->assertInstanceOf('\Geocoder\Model\Address', $result);
         $this->assertEquals(43.6155351, $result->getCoordinates()->getLatitude(), '', 0.01);
         $this->assertEquals(1.4525647, $result->getCoordinates()->getLongitude(), '', 0.01);
-        $this->assertTrue($result->getBounds()->isDefined());
+        $this->assertNotNull($result->getBounds());
         $this->assertEquals(43.6154556274414, $result->getBounds()->getSouth(), '', 0.01);
         $this->assertEquals(1.4524964094162, $result->getBounds()->getWest(), '', 0.01);
         $this->assertEquals(43.6156005859375, $result->getBounds()->getNorth(), '', 0.01);