+++ tests/Geocoder/Tests/Provider/YandexTest.php
@@ -100,7 +100,7 @@ class YandexTest extends TestCase
         $this->assertInstanceOf('Geocoder\Model\Address', $result);
         $this->assertEquals(48.863277, $result->getCoordinates()->getLatitude(), '', 0.01);
         $this->assertEquals(2.389016, $result->getCoordinates()->getLongitude(), '', 0.01);
-        $this->assertTrue($result->getBounds()->isDefined());
+        $this->assertNotNull($result->getBounds());
         $this->assertEquals(48.861926, $result->getBounds()->getSouth(), '', 0.01);
         $this->assertEquals(2.386967, $result->getBounds()->getWest(), '', 0.01);
         $this->assertEquals(48.864629, $result->getBounds()->getNorth(), '', 0.01);
@@ -135,7 +135,7 @@ class YandexTest extends TestCase
         $this->assertInstanceOf('Geocoder\Model\Address', $result);
         $this->assertEquals(55.675676, $result->getCoordinates()->getLatitude(), '', 0.01);
         $this->assertEquals(12.567593, $result->getCoordinates()->getLongitude(), '', 0.01);
-        $this->assertTrue($result->getBounds()->isDefined());
+        $this->assertNotNull($result->getBounds());
         $this->assertEquals(55.614999, $result->getBounds()->getSouth(), '', 0.01);
         $this->assertEquals(12.45295, $result->getBounds()->getWest(), '', 0.01);
         $this->assertEquals(55.73259, $result->getBounds()->getNorth(), '', 0.01);
@@ -192,7 +192,7 @@ class YandexTest extends TestCase
         $this->assertInstanceOf('Geocoder\Model\Address', $result);
         $this->assertEquals(38.897695, $result->getCoordinates()->getLatitude(), '', 0.01);
         $this->assertEquals(-77.038692, $result->getCoordinates()->getLongitude(), '', 0.01);
-        $this->assertTrue($result->getBounds()->isDefined());
+        $this->assertNotNull($result->getBounds());
         $this->assertEquals(38.891265, $result->getBounds()->getSouth(), '', 0.01);
         $this->assertEquals(-77.046921, $result->getBounds()->getWest(), '', 0.01);
         $this->assertEquals(38.904125, $result->getBounds()->getNorth(), '', 0.01);
@@ -226,7 +226,7 @@ class YandexTest extends TestCase
         $this->assertInstanceOf('Geocoder\Model\Address', $result);
         $this->assertEquals(53.898077, $result->getCoordinates()->getLatitude(), '', 0.01);
         $this->assertEquals(27.563673, $result->getCoordinates()->getLongitude(), '', 0.01);
-        $this->assertTrue($result->getBounds()->isDefined());
+        $this->assertNotNull($result->getBounds());
         $this->assertEquals(53.896867, $result->getBounds()->getSouth(), '', 0.01);
         $this->assertEquals(27.561624, $result->getBounds()->getWest(), '', 0.01);
         $this->assertEquals(53.899286, $result->getBounds()->getNorth(), '', 0.01);
@@ -288,7 +288,7 @@ class YandexTest extends TestCase
         $this->assertInstanceOf('Geocoder\Model\Address', $result);
         $this->assertEquals(48.863212, $result->getCoordinates()->getLatitude(), '', 0.01);
         $this->assertEquals(2.388773, $result->getCoordinates()->getLongitude(), '', 0.01);
-        $this->assertTrue($result->getBounds()->isDefined());
+        $this->assertNotNull($result->getBounds());
         $this->assertEquals(48.86294, $result->getBounds()->getSouth(), '', 0.01);
         $this->assertEquals(2.387497, $result->getBounds()->getWest(), '', 0.01);
         $this->assertEquals(48.877038, $result->getBounds()->getNorth(), '', 0.01);
@@ -335,7 +335,7 @@ class YandexTest extends TestCase
         $this->assertInstanceOf('Geocoder\Model\Address', $result);
         $this->assertEquals(48.87132, $result->getCoordinates()->getLatitude(), '', 0.01);
         $this->assertEquals(2.404017, $result->getCoordinates()->getLongitude(), '', 0.01);
-        $this->assertTrue($result->getBounds()->isDefined());
+        $this->assertNotNull($result->getBounds());
         $this->assertEquals(48.86294, $result->getBounds()->getSouth(), '', 0.01);
         $this->assertEquals(2.387497, $result->getBounds()->getWest(), '', 0.01);
         $this->assertEquals(48.877038, $result->getBounds()->getNorth(), '', 0.01);
@@ -394,7 +394,7 @@ class YandexTest extends TestCase
         $this->assertInstanceOf('Geocoder\Model\Address', $result);
         $this->assertEquals(60.454462, $result->getCoordinates()->getLatitude(), '', 0.01);
         $this->assertEquals(22.256561, $result->getCoordinates()->getLongitude(), '', 0.01);
-        $this->assertTrue($result->getBounds()->isDefined());
+        $this->assertNotNull($result->getBounds());
         $this->assertEquals(60.45345, $result->getBounds()->getSouth(), '', 0.01);
         $this->assertEquals(22.254513, $result->getBounds()->getWest(), '', 0.01);
         $this->assertEquals(60.455474, $result->getBounds()->getNorth(), '', 0.01);
@@ -429,7 +429,7 @@ class YandexTest extends TestCase
         $this->assertInstanceOf('Geocoder\Model\Address', $result);
         $this->assertEquals(40.874651, $result->getCoordinates()->getLatitude(), '', 0.01);
         $this->assertEquals(29.129562, $result->getCoordinates()->getLongitude(), '', 0.01);
-        $this->assertTrue($result->getBounds()->isDefined());
+        $this->assertNotNull($result->getBounds());
         $this->assertEquals(40.860413, $result->getBounds()->getSouth(), '', 0.01);
         $this->assertEquals(29.107230, $result->getBounds()->getWest(), '', 0.01);
         $this->assertEquals(40.876111, $result->getBounds()->getNorth(), '', 0.01);